// src/models/TextType.ts
enum TextType {
    h1 = "h1",
    h2 = "h2",
    h3 = "h3",
    h4 = "h4",
    quote = "quote",
    question = "question",
    bodyLarge = "body-large",
    body = "body",
    bodyCompact = "body-compact",
    bodySmall = "body-small",
    bodyXSmall = "body-x-small",
    buttons = "buttons",
    buttonsSmall = "buttonsSmall",
}

export default TextType;